+++ client/src/Components/Home/Home.jsx
@@ -1,5 +1,5 @@
 // import React from "react";
-import NavBar from "./NavBar";
+import NavBar from "../NavBar/NavBar";
 
 
 export default function Home() {
